+++ app/__init__.py
@@ -6,14 +6,18 @@ import os
 
 from flask import Flask, jsonify
 from flask_sqlalchemy import SQLAlchemy
+from flask_migrate import Migrate
+from flask_bcrypt import Bcrypt
 
 
 from app.config import app_config
 from app.api.views import users_blueprint
 
 
-# Instantiate the DB
+# Instantiate extensions
 db = SQLAlchemy()
+migrate = Migrate()
+bcrypt = Bcrypt()
 
 
 def create_app():
@@ -27,6 +31,8 @@ def create_app():
 
     # Set up extensions
     db.init_app(app)
+    bcrypt.init_app(app)
+    migrate.init_app(app, db)
 
     # Register blueprints
     app.register_blueprint(users_blueprint)