+++ app/api/models.py
@@ -3,7 +3,9 @@
 
 import datetime
 
-from app import db
+from flask import current_app
+
+from app import db, bcrypt
 
 
 # User model
@@ -14,13 +16,20 @@ class User(db.Model):
     __tablename__ = "users"
     # pylint: disable=invalid-name
     id = db.Column(db.Integer, primary_key=True, autoincrement=True)
-    username = db.Column(db.String(128), nullable=False)
-    email = db.Column(db.String(128), nullable=False)
-    active = db.Column(db.Boolean(), default=False, nullable=False)
+    username = db.Column(db.String(128), unique=True, nullable=False)
+    email = db.Column(db.String(128), unique=True, nullable=False)
+    password = db.Column(db.String(255), nullable=False)
+    active = db.Column(db.Boolean, default=True, nullable=False)
+    admin = db.Column(db.Boolean, default=False, nullable=False)
     created_at = db.Column(db.DateTime, nullable=False)
 
-    def __init__(self, username, email):
+    def __init__(
+            self, username, email, password,
+            created_at=datetime.datetime.utcnow()):
         """Initialize MetaGenScope User model."""
         self.username = username
         self.email = email
-        self.created_at = datetime.datetime.utcnow()
+        self.password = bcrypt.generate_password_hash(
+            password, current_app.config.get('BCRYPT_LOG_ROUNDS')
+        ).decode()
+        self.created_at = created_at