+++ app/config.py
@@ -12,6 +12,7 @@ class Config(object):
     TESTING = False
     SQLALCHEMY_TRACK_MODIFICATIONS = False
     SECRET_KEY = os.environ.get('SECRET_KEY')
+    BCRYPT_LOG_ROUNDS = 13
 
 
 class DevelopmentConfig(Config):
@@ -20,6 +21,7 @@ class DevelopmentConfig(Config):
     DEBUG = True
     SECRET_KEY = 'my_precious'
     SQLALCHEMY_DATABASE_URI = os.environ.get('DATABASE_URL')
+    BCRYPT_LOG_ROUNDS = 4
 
 
 class TestingConfig(Config):
@@ -29,6 +31,7 @@ class TestingConfig(Config):
     TESTING = True
     SECRET_KEY = 'my_precious'
     SQLALCHEMY_DATABASE_URI = os.environ.get('DATABASE_TEST_URL')
+    BCRYPT_LOG_ROUNDS = 4
 
 
 class StagingConfig(Config):