+++ app/tests/test_config.py
@@ -29,6 +29,7 @@ class TestDevelopmentConfig(TestCase):
             app.config['SQLALCHEMY_DATABASE_URI'] ==
             os.environ.get('DATABASE_URL')
         )
+        self.assertTrue(app.config['BCRYPT_LOG_ROUNDS'] == 4)
 
 
 class TestTestingConfig(TestCase):
@@ -48,6 +49,7 @@ class TestTestingConfig(TestCase):
             app.config['SQLALCHEMY_DATABASE_URI'] ==
             os.environ.get('DATABASE_TEST_URL')
         )
+        self.assertTrue(app.config['BCRYPT_LOG_ROUNDS'] == 4)
 
 
 class TestProductionConfig(TestCase):
@@ -61,6 +63,7 @@ class TestProductionConfig(TestCase):
         """Ensure appropriate configuration values for production environment."""
         self.assertFalse(app.config['DEBUG'])
         self.assertFalse(app.config['TESTING'])
+        self.assertTrue(app.config['BCRYPT_LOG_ROUNDS'] == 13)
 
 
 if __name__ == '__main__':