+++ app/tests/utils.py
@@ -1,0 +1,19 @@
+"""Common utility methods for use in testing."""
+
+import datetime
+
+
+from app import db
+from app.api.models import User
+
+
+def add_user(username, email, password, created_at=datetime.datetime.utcnow()):
+    """Wrap functionality for adding user."""
+    user = User(
+        username=username,
+        email=email,
+        password=password,
+        created_at=created_at)
+    db.session.add(user)
+    db.session.commit()
+    return user